Dispatch for default/CSR array
