+++ python/mxnet/__init__.py
@@ -106,3 +106,5 @@ from . import tvmop
 from . import numpy_op_signature
 from . import numpy_dispatch_protocol
 from . import numpy_op_fallback
+
+from . import _global_var
