+++ python/mxnet/ndarray/_internal.py
@@ -23,15 +23,18 @@ import sys as _sys
 try:
     if int(_os.environ.get("MXNET_ENABLE_CYTHON", True)) == 0:
         from .._ctypes.ndarray import NDArrayBase, CachedOp
-        from .._ctypes.ndarray import _set_ndarray_class, _imperative_invoke, _set_np_ndarray_class
+        from .._ctypes.ndarray import _imperative_invoke
+        from .._global_var import _set_ndarray_class, _set_np_ndarray_class
     else:
         from .._cy3.ndarray import NDArrayBase, CachedOp
-        from .._cy3.ndarray import _set_ndarray_class, _imperative_invoke, _set_np_ndarray_class
+        from .._cy3.ndarray import _imperative_invoke
+        from .._global_var import _set_ndarray_class, _set_np_ndarray_class
 except ImportError:
     if int(_os.environ.get("MXNET_ENFORCE_CYTHON", False)) != 0:
         raise ImportError("Cython Module cannot be loaded but MXNET_ENFORCE_CYTHON=1")
     from .._ctypes.ndarray import NDArrayBase, CachedOp
-    from .._ctypes.ndarray import _set_ndarray_class, _imperative_invoke, _set_np_ndarray_class
+    from .._ctypes.ndarray import _imperative_invoke
+    from .._global_var import _set_ndarray_class, _set_np_ndarray_class
 
 from ..base import _Null
 try:
